nft/core/src/main/kotlin/com/rarible/protocol/nft/core/service/item/reduce/forward/ForwardValueItemReducer.kt
package com.rarible.protocol.nft.core.service.item.reduce.forward

import com.rarible.core.entity.reducer.service.Reducer
import com.rarible.protocol.nft.core.model.Item
import com.rarible.protocol.nft.core.model.ItemEvent
import org.springframework.stereotype.Component

@Component
class ForwardValueItemReducer : Reducer<ItemEvent, Item> {
    override suspend fun reduce(entity: Item, event: ItemEvent): Item {
        return when (event) {
            is ItemEvent.ItemMintEvent -> entity.copy(supply = entity.supply + event.supply)
            is ItemEvent.ItemBurnEvent -> entity.copy(supply = entity.supply - event.supply)
            is ItemEvent.ItemCreatorsEvent,
            is ItemEvent.ItemTransferEvent-> entity
            is ItemEvent.LazyItemBurnEvent, is ItemEvent.LazyItemMintEvent ->
                throw IllegalArgumentException("This events can't be in this reducer")
        }
    }
}

